+++ src/common/const/constants.js
@@ -3,3 +3,7 @@ export const NO_ID = 'noId';
 
 export const CREATE_UNITS_PERM = 'finance.acquisitions-units-assignments.assign';
 export const MANAGE_UNITS_PERM = 'finance.acquisitions-units-assignments.manage';
+
+export const OVERALL_ROLLOVER_STATUS = {
+  inProgress: 'In Progress',
+};